requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-openai==0.3.0
langgraph==0.2.62
langgraph-checkpoint-sqlite==2.0.6
chromadb==0.5.23
pandas==2.2.3
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
